create preference parser
