+++ .gitignore
@@ -388,10 +388,23 @@ MigrationBackup/
 # End of https://www.gitignore.io/api/c++,visualstudio
 
 
+#####################################################
+## Visual Studio Code
+#####################################################
+settings.json
+
+
+#####################################################
+## Visual Studio
+#####################################################
 *.sln
 *.vcxproj
 *.vcxproj.filters
 
+
+#####################################################
+## other
+#####################################################
 main.cpp
 
 test/
